+++ tests/data/hilti_exp4_small/lidar/constraints.txt
@@ -1,0 +1,3 @@
+1 2 0.999997010223193 0.000262387847156579 -0.00243119256614615 -0.00197101916043594 -0.000290858563825453 0.999931303638918 -0.0117176535308903 -0.00313535578160279 0.00242795098218007 0.0117183256308997 0.99992839038525 -0.00461651519089804 3.66063230799257e-05 9.9545265818141e-06 5.45778023651052e-06 1.17304954869707e-05 -7.80291236405833e-06 4.46945576709257e-06 9.9545265818141e-06 4.71070688195386e-05 -4.93945328139692e-06 2.84701686145373e-06 -9.59346068897752e-06 5.82413595693807e-06 5.45778023651052e-06 -4.93945328139692e-06 1.91908313282482e-05 5.11299277559985e-06 2.28667091945577e-06 -1.3876334784889e-08 1.17304954869707e-05 2.84701686145373e-06 5.11299277559985e-06 5.73720097472749e-05 -1.66078193514253e-07 5.82114183966509e-06 -7.80291236405833e-06 -9.59346068897752e-06 2.28667091945577e-06 -1.66078193514253e-07 3.38316227377772e-05 -6.12267454951798e-06 4.46945576709257e-06 5.82413595693807e-06 -1.3876334784889e-08 5.82114183966509e-06 -6.12267454951798e-06 3.35653839569578e-05 2210 2160 0 32 80 2160 2156 0 15 62 0 0 0 0 0 32 15 0 2946 0 80 62 0 0 2581
+0 2 0.999999607793699 0.000372502409736026 -0.000803526229839881 -0.00103627901334157 -0.000381185264649734 0.999941248803592 -0.0108329884569136 -0.00482156465398071 0.000799443707407689 0.0108332905005058 0.999940998612763 -0.00212086629720318 2.72454562307393e-05 2.66341586431979e-06 5.38122051750532e-07 2.85467343500581e-06 1.94569838520849e-06 8.75417584765239e-07 2.66341586431979e-06 3.96461911496845e-05 -3.81514933573987e-06 3.73281351458948e-06 -5.60295626600072e-07 -3.94680424596079e-07 5.38122051750532e-07 -3.81514933573987e-06 2.01396798293104e-05 3.44579243056545e-06 -1.97235491741986e-06 -1.36782134686471e-06 2.85467343500581e-06 3.73281351458948e-06 3.44579243056545e-06 6.03462195775359e-05 2.52203279791649e-06 -1.52163701596561e-06 1.94569838520849e-06 -5.60295626600072e-07 -1.97235491741986e-06 2.52203279791649e-06 3.28939298464002e-05 -7.08018568077332e-06 8.75417584765239e-07 -3.94680424596079e-07 -1.36782134686471e-06 -1.52163701596561e-06 -7.08018568077332e-06 3.05374537500662e-05 2248 2194 0 0 65 2194 2198 0 0 32 0 0 0 0 0 0 0 0 2947 0 65 32 0 0 1813
+0 1 0.999998910655187 -1.18232921565255e-05 -0.0014759907350704 -0.00238071336084558 1.55767607006661e-05 0.999996766370788 0.00254303073768344 -0.00106469847713755 0.00147595589526827 -0.00254305095860057 0.999995677213665 -0.000506407620746786 5.12425651759002e-05 1.38264914278017e-05 7.37969925898122e-06 1.30243155201676e-05 -5.76734076473127e-06 -4.92645632408824e-06 1.38264914278017e-05 3.60197184898574e-05 -2.03387990175962e-06 1.76746285113647e-06 -4.64405922450097e-06 -5.23604957886322e-06 7.37969925898122e-06 -2.03387990175962e-06 2.02796437701656e-05 6.61143649645025e-06 -2.36543550071326e-07 -3.27487375186265e-06 1.30243155201676e-05 1.76746285113647e-06 6.61143649645025e-06 5.44100781291303e-05 4.05171274887978e-07 2.61698398817441e-06 -5.76734076473127e-06 -4.64405922450097e-06 -2.36543550071326e-07 4.05171274887978e-07 3.51832901766832e-05 -5.03862118594466e-06 -4.92645632408824e-06 -5.23604957886322e-06 -3.27487375186265e-06 2.61698398817441e-06 -5.03862118594466e-06 3.61794406580814e-05 2159 2120 0 24 70 2120 2119 0 17 59 0 0 0 0 0 24 17 0 2874 0 70 59 0 0 1799
